A bunch of failing section 8 tests
