+++ crates/httpwg/src/rfc9113/mod.rs
@@ -30,3 +30,4 @@ pub mod _4_http_frames;
 pub mod _5_streams_and_multiplexing;
 pub mod _6_frame_definitions;
 pub mod _7_error_codes;
+pub mod _8_expressing_http_semantics_in_http2;
